NHRP: added fields for ipv6 addresses
Added fields for IPv6 addresses to packet-nhrp.c that clone fields for IPv4
addresses. Dissection is done with check for length 16. Additionaly changed
dissection for Extension data: if length of extension is less than source
length + 4 the source length is not added to tree — the whole remainder is
considered as data.