ssh: Adjust length warning for modes that do not encrypt packet length
RFC 4253 6: "[T]he length of the concatenation of 'packet_lVength',
'padding_length', 'payload', and 'random padding' MUST be `a multiple
of the cipher block size or 8, whichever is larger,... even when
using stream ciphers."

Modes that do not encrypt plen with the same key as the other three
cannot follow this as written and delete 'packet_length' from the
above sentence. As padding_length is one byte and random_padding at
least four, packet_length must be at least 8 in all modes.

This is not really necessary for AES GCM, because the block size is
16, but does affect some other modes such as -etm modes or ChaCha20.