ssh: Parse env and exec CHANNEL_REQUEST messages
Note that unfortunately the SSH protocol does not specify the encoding
for these strings, which must match whatever the server expects. (The
bytes returned in CHANNEL_DATA messages are affected by environment
variables and the shell defaults, but these strings are passed directly
to execve.) In practice they must not have internal NULs or use wide
characters (so no NUTF-16), and OpenSSH ports to Windows, IBM z/OS, etc.
force the use of UTF-8 or ISO-8859-1 rather than attempt to deal with
non ASCII-compatible encodings.

Also fix the indentation in ssh_dissect_connection_specific to match the
rest of the dissector and EditorConfig.