epan: register_tap_listener could cause double free on error
register_tap_listener has various paths where an error string can be
returned, indicating an error while setting up the tap listener. In
one of the paths, where the display filter is processed, a call is made
to free_tap_listener to release all of the tap listener structure.
Included in this is an attempt to call the tap.finish callback.

This has two problems associated with it. A calling function can not
assume the tap.finish callback has been called on error, since not all
error paths make this call. Due to this fact non of the callers of
register_tap_listener make this assumption, on error they perform
cleanup operations themselves. This could cause double free errors.

Secondly, since the tap listener struct has not been setup completely
when free_tap_listener is called in the error path, the call to the
tap.finish callback is never made. This is what the comment is about.

As a matter of fact non of the tap listener struct parts on which
free_tap_listener operates has been setup yet, apart from the struct
itself. Therefore the whole call to free_tap_listener in this error
path is not better than simply freeing the tap listener struct itself.

Therefor, to simplify the code and avoid any future double free errors,
swap the use of free_tap_listener by a simple free of the struct.