Gryphon: Fix retrieval of destination ChannelID in CMD_MSGRESP_MODIFY
The destination channel ID is found at offset - 9, not offset - 5;
using offset - 5 is presumably from forgetting about the 4 byte
command header between the Command Data and the Frame Header:

https://www.dgtech.com/product/gryphon/manual/html/GCprotocol/client_howto/GCframeformat.php#Commands
https://www.dgtech.com/product/gryphon/manual/html/GCprotocol/commands/Responder/modify_response.php

Note that cmd_sched already correctly uses offset - 9 to retrieve the
destination channel ID in an analogous situation.

Also, do not use the %c printf specifier to print a byte, as that is
not guaranteed to be valid UTF-8 in fuzzed data.

Fix #20585