Qt: Wildcarded disconnect considered harmful
Only disconnect the main stack from the signal that we connected,
instead of doing a wildcarded disconnect all signals. In Qt 6.8.4,
6.9.1, and later, the wildcard disconnect causes a Qt warning about
disconnecting the destroyed() signal. (Cf. #20174 and commit
7a7b99e0d96ff378f531b40300ce41b65a2c049a for a similar case,
which caused crashing in QAccessibilityCache, i.e.
https://bugreports.qt.io/browse/QTBUG-134610 )

Prevents warning:
** (wireshark:1860454) 08:10:05.984811 [GUI WARNING] -- QObject::disconnect: wildcard call disconnects from destroyed signal of QStackedWidget::mainStack

This is the only signal that needs to be disconnected, as it's the
only one we manually connect (and the one mentioned in commit
a8a3903e55af23bdce550a95ffe14a398fd1204c which introduced the
disconnect.)