mq: Don't get strings to append to COL_INFO if pinfo is NULL
There's an attempt at optimization in the mq dissector that appends
some strings to COL_INFO only if pinfo is NULL. Don't bother extracting
the strings if pinfo is NULL, because they're not otherwise used.

This also prevents a segfault introduced when switching to pinfo->pool
scoped memory in that case.

Fix #20593