GSM RP: Dissect Originator Address as a Calling Party BCD Number
3GPP TS 24.011 8.2.5.1 and 8.2.5.2 mention that octets 3, 4, etc.
of both the RP-Originator Address and RP-Destination Address are
the same as those defined for the Called Party BCD Number IE in
3GPP TS 24.008. In that spec, 10.5.4.7 and 10.5.4.9 shows that
Called Party BCD Number and Calling Party BCD Number have the
same structure as well, except that in the Called Party BCD Number
an extension bit is always 1, whereas in Calling Party the extension
bit might be zero (indicating an extra octet.)

Indeed, inside packet-gsm_a_dtap, both Called and Calling party are
dissected the same, calling the same helper function with a different
hfid. (That is, even though the Called Party never has an extension,
if the extension bit wrong is zero, the extension octet will be added.)

So dissect the Originator Address with Calling Number, which will
have no change other than the GSM RP Originator Address will have
a different hfid used for its digits than the Destination Address,
so that they can be filtered separately.
It is less confusing to have the Originator Address referred to as
a Calling Address in the tree than a Called Address.