epan: Pull postdissector priming internal to libwireshark
If postdissector want hfids on the first pass, the epan_dissect_t
needs to be primed. We can do that when dissecting, as we know
whether or not the frame has been visited and whether postdissectors
were registered and want hfids. This keeps applications from having
to remember to call the function to prime.

Ping #20600, as this makes it easier to deal with that later by
centralizing the logic.