Qt: Fix crash reloading Lua plugins with non default non custom columns
Since register columns now get _ws.col registered fields, we have to
recreate the columns even when there aren't custom columns, as one
of those might have changed (as it does when reloading Lua plugins.)

We don't have to if pref.col_list is NULL and stays NULL (the columns
aren't registered and deregistered in that case), and if it switches
to or from NULL, we recreate the columns anyway.

In the case where the column preference was not the default, but it
didn't contain any custom columns, we weren't rebuilding the columns.

Fixup 1b82eda9ebb72a4ac5198a44072f8694ffc1e060

Fix #20588