MATE: Fix SCS handling of huge strings
The Single Copy Strings (reference counted strings) in MATE truncates
huge strings to a maximum value before inserting them. However,
subsequent lookups for additional subscription or unsubscription
don't truncate to that maximum length, so it doesn't work.

Change the hash and comparison function to use the maximum length.
Automatically free the strings when they are removed from the hash table
(including when the hash table is destroyed, though that doesn't happen yet.)

While we're at it, simplify the code by removing the slice allocator,
since on modern GLib (>= 2.76) the GSlice allocator was removed as
modern platforms have more efficient default allocators (like slab
allocators), so this is just g_malloc and g_free anyway. Even with
older GLib recent platforms are probably as faster or faster using
the default allocators.
 (https://docs.gtk.org/glib/memory-slices.html)