MATE: Simplify hash table calls
g_hash_table_foreach_remove() with a function that just returns true
is the same thing as g_hash_table_remove_all, but more complicated.
There's also no point in destroying hash tables and recreating them
exactly the same. (There could be a point in reducing a reference
count and changing the pointer, but that's not what's happening here.)