MATE: Free runtime created PDUs
Free the runtime created PDUs instead of having them leak.

Store the PDUs as a GPtrArray in the hash table instead of having
each pdu have a next pointer (so that freeing them doesn't require
recursion; GLists don't store in hash tables because the head isn't
guaranteed to be stable.)

Store the map of the last id for each pducfg in the runtime data,
rather than altering the parsed mate_config and resetting it.