wifi-nan: fix the AMBTT appearance and Discovery Beacon info
The AMBTT (Anchor Master Beacon Transmission Time) is
the lower 4 octets of the Anchor Master TSF.
Since the TSF is presented in little endian, it is
confusing to see the AMBTT in big endian.

In addition, better detect and present a Discovery Beacon.
According to NAN spec (3.3.8.1), a Discovery Beacon interval
can be somewhere between 50 to 200 TU (Time Units).
Hence, Detect a Discovery Beacon if the interval is not 512 TU,
and not only if the interval is 100 TU.