Qt: Draw the coloring rules checkbox with the application palette
On at least some platforms (X11/Wayland/Windows, but not Mac, AIUI),
Qt Item Views use the text color/Foreground color to draw the check
mark in a checkbox, but always use the same theme/palette color for
the square behind the checkbox. (The background color is used behind
the checkbox.) That doesn't look good for certain foreground colors.

Draw the item check area a second time with the standard colors.

Fix #20603