capture_opts: set the timestamp type to a *copy* of the default.
Timestamp types for interfaces are strings, not the reuslt of calling
pcap_tstamp_type_name_to_val() on the string.  If the timestamp type for
an interface is changed, the old timestamp type is freed and replaced by
a copy of the new type; this means that, when we create a set of options
for a new interface, we must set its timestamp type to a *copy* of the
default, so that, if it's subsequently freed, we're not freeing it out
from under the default or from other interfaces that have been assigned
the default.

That's done with other string options, so it should be done here as
well.

This fixes a crash reported on the security mailing list.