Allow zero length source lengths to tvb_bytes_to_str[_punct]
Move the "source length is zero" check in bytes_to_str and
bytes_to_str_punct before the "was a null pointer passed in
check." The input isn't a null terminated string, only an array
of bytes, and ensure_contiguous returns a NULL pointer for length
zero.

Change the assertions in the tvb_ versions to allow the length to
be zero as well, to reduce extra checking that dissectors might
need to do.

Fix #20607 (See the related #19705).