file: Work around DSBs not at the start of a file
If we encounter a DSB not at the start of a file when reading it,
mark the file as having unsaved changes. After all, if we save the
file, we'll move the DSBs to the beginning.

Also, trigger a redissection after we're done. A DSB in the middle
might have come slightly too later for packets, and it might change
dissection. (Should this be under a pref?)

For live capture we at least mark the unsaved changes but do not
queue redissection yet, which might also work but needs more testing.

Fix #20558