DTLS: Consider epoch in DTLS 1.2 reassembly of rehandshake
In DTLS 1.2, the message_seq is reset and the epoch incremented
at each rehandshake (i.e., renegotiation), so concatenate them
for a fragment sequence number. However, in DTLS 1.3 the
message_seq is not reset in a post-handshake message exchange
(so having passed in 0 as the epoch here instead of the real
64-bit DTLS 1.3 epoch should work unless message_seq has
wrapped around, which is unlikely as it is only incremented
in handshake messages).

        https://www.rfc-editor.org/rfc/rfc9147.html#section-5.2-6
        https://datatracker.ietf.org/doc/html/rfc6347#section-4.1

Also use proto_add_item_ret_uint in a few places nearby.

Fix #20362 for a greater variety of cases (e.g., when the server
does not send a HelloRequest, or when the client changes the cert
or other handshake information on the rehandshake.)