pcapng: Deal with overflow with very fine timestamp resolution
If the time resolution in a pcapng is 10^-N where N > 10 or
2^-N where N > 32, then the nanoseconds calculations can overflow.
If it does overflow, we can calculate the nanoseconds with exact
integer arithmetic one of two different ways without overflow,
depending on whether it's a power of 10 or a power of 2.

Alternatively, we could use 128 (or 96) bit integers on platforms
that support that, such as C23 or gcc, clang, and ICC on most 64-bit
platforms. It's not clear that always doing so would be faster, though.

(Note that in such cases, we can't write any times past
2001-09-09T01:46:46 without using a time offset, so if we're not
writing one, we have to reduce precision. It's not clear that works
properly.)

Fix #20608 (but check whether this file gets written correctly on export.)