wiretap: Consolidate limiting IDB precision to nanoseconds
Currently, nstime_t does not store more than nanoseconds, which means
that wtap_rec and frame_data are always reduced to nanosecond precision
after reading from a file, even if the file had greater precision.
That means that wiretap can't really write greater precision to a file
(outside of saving that is really just moving or copying a file at the
binary level), so don't claim that we have greater precision that we
have.

There are some hacks to try to limit the precision, but it's not done
in every code path, and so the picosecond resolution file from #20608
isn't written correctly when exporting packets to a pcapng (even if the
timestamp was correctly converted to nanosecond resolution for display.)
It *is* written correctly for other file formats that don't support
IDBs.

Consolidate all the "altering an IDB so it doesn't have a TSOFFSET
option or a TS resolution greater than nanoseconds" into one place,
where an IDB is added for dumping, instead of small hacks in various
places, included the pcapng writer simply not writing a TSOFFSET if
it's there or trying to remove it on read.

In the future, nstime_t should have greater precision (the "ns" part
will then be a misnomer.) For example, since time_t is already 64
bits on most platforms, and most platforms are 64-bit aligned, the
fractional part could be expanded considerably without using extra
space. Some space could even be left over to indicate the stored
precision (see, e.g., #15579).