proto: Add check for valid value_pos and name_pos
Check if value_pos is after end of the buffer and set it to end
of buffer if it is. In this case the buffer has no value to show,
but this is a corner case and should be ok.

Check if name_pos is at the end of the buffer to be truncated,
and put trunc_str at the beginning if it's no room at the end.

Fixes #20613