http-urlencoded: Handle byte sequences with no '='
The "aberrant monstrosity" that is the application/x-www-form-urlencoded
format is supposed to be serialized with a '=' when there is a name with
an empty string value, but parsing is supposed to handle the case where
there is no '=' separator by treating it entirely as name.

Always skip on where the '&' was found when finishing a name-value
pair byte sequence to handle both cases with a minimum of tests.

https://url.spec.whatwg.org/#application/x-www-form-urlencoded
https://url.spec.whatwg.org/#urlencoded-parsing
https://url.spec.whatwg.org/#urlencoded-serializing

Fix #20615