Use ws_log for busmaster and peak_trc
Use the ws_log system instead of printf() for reporting messages
from the busmaster and peak_trc wiretaps. If Wireshark
isn't compiled for a Debug target, the compiler will optimize away the
calls to log debug messages, and the rest will be presented in a
consistent way to the user.

Ping #19519