DNP3: Fix Y2038 issue with timestamps on LLP64 systems (e.g., Windows)
Cast to a time_t, not a long, when storing in the nstime_t secs, so as
to work on platforms where time_t is 64-bit but long is 32 bit, such
as Windows.

These should just be added with ENC_TIME_MSECS|ENC_LITTLE_ENDIAN, which
will be changed in a later commit, but this one liner is easier to
backport.

Fix #20618