merge: Treat no tsresol option as 6 when comparing IDBs
An IDB that lacks the if_tsresol option is treated as having a
timestamp resolution of 6 (microseconds). So if one IDB has a
non-6 tsresol option value and the other has no tsresol option,
those should be treated as different for merging purposes.
Conversely, if one IDB has an explicit tsresol option of 6 and
the other has no tsresol option, but they are otherwise identical,
they are the same interface. (Some libraries, including wiretap,
might remove a redundant TSRESOL of 6.)