wmem: Cache full hash in wmem_map
Add the entire unmasked hash to each item stored in wmem_map.
This increases the size and memory allocation, but it eliminates
having to rehash when growing the table, and can replace possibly
expensive equality tests with comparisons of already computed
hash values in the case of mask hash collisions.

This reduces the instruction count on a x86_64 machine by roughly
1% for TShark processing an empty file.