proto: Use wmem_map for gpa_name_map
Use a wmem_map for gpa_name_map, because we can reserve an initial
number of entries. GHashTable doesn't have that functionality, and
the large number of resizes up to 245760 or 491510 elements impose
a startup cost.

Note that GHashTable is an open addressing / probing hash table,
whereas wmem_map is a chaining hash table.