wmem_map: Add a function to reserve a minimum capacity
Reallocation and copying of existing elements when the table size
grows can be expensive for very large tables, so add a function
to reserve a minimum capacity.

Also fix a few types; the number of buckets is limited to 2^32, but
the actual number of items stored can actually exceed that (since this
is a chained hash table). Correcting the return value of wmem_map_size
would be an ABI break, though, and maybe isn't worth it.