epan: Avoid hashing twice when checking for duplicate proto names
Ever since GLib 2.40 (we've required 2.50 since Wireshark 4.0),
g_hash_table_insert has returned a boolean, which is FALSE if the
key already existed in the hash table, indicating a sort of failure
condition. That means we don't need to do hash table lookups before
inserting. Just try the insert, and if it returns false, report the
bug. As we cause the program to halt on startup, we don't have to
worry about cleanup.

This avoids doing extra string hashes on each of the name, short name,
and filter name.