bluetooth: Put generated data in a different file
Make it easier to update the Bluetooth UUIDs and other information by
writing them to a separate file. Pregenerate the name, short name,
and filter abbreviations for the BT ATT handles too instead of creating
the strings dynamically at handoff time.

Eventually we should run this automatically once a week with the
other scripts.