addr_resolv: Fix service lookups of same port number, different types
Store in the cache with both port number and type, instead of storing
a struct that has the four port types with services and the numeric
string. (Store the numeric strings under PT_NONE, noting that we don't
return it for try_serv_name.)

This prevents confusion when looking up the same port number for a
different port type. Previously, the other port type might have NULL
stored because we only looked in the custom or services table for a
different type. That made it so we had to choose between either failing
to lookup for a different port type, or repeatedly looking up port
numbers that have no service name. (Or, I suppose, always looking up
and storing results for all 4 port used port types the first time a
number was requested.)

Since the port number and port type are both 16 bit, these could be
stored in the table as a single 32 bit integer with GUINT_TO_POINTER
directly, but a key type struct might be simpler since the table is
exposed to the Resolved Addresses dialog.

Fixup 73b5d127c6f7550913aac76b98c8fd533defa0b5