Make the XML DTDs more spec compliant
Processing Instructions in XML are defined such that the PITarget must
follow immediately after the initial '<?', i.e. there can be no
whitespace. The PITarget is a Name, which cannot have ":" as a character
other than the initial character.

  https://www.w3.org/TR/xml/#NT-PI
  https://www.w3.org/TR/xml/#NT-Name

If we later want to parse the DTDs with a compliant XML and DTD parser
instead of a hand-rolled flex and lemon based parser, they should be
made more compliant. Change the DTDs and the Lemon parser accordingly.