netlink: fix selected bytes for padding
At that point of the code, offset points to the beginning of the payload
of the attribute. The padding starts at the signalled length of the
attribute (which includes the header size), so the padding starts at
offset + signalled length - header size (4).

Also attaching the padding to the attribute tree instead of the global
tree makes more sense.