MQ: fix the code that handles the MQCONN reply.
Add comments indicating what the code in question is doing.

The application type should only be used to guess the byte order; it's
*not* the character encoding for th application, so don't use it for
that. Make it local to the code that handles the MQCONN reply, as that's
the only place it's used. Get rid of the extra assignment to it outside
that code.