epan: Use a hash table for descriptions of handles in Decode As tables
Decode As tables require handles have descriptions, and that they be
unique for displaying and selecting in the GUI, and for writing to
the decode_as_prefs UAT. They are frequently looked up by the
description. Therefore, use a hash table.

The list is used in a handful of other situations, notably for error
messages from tshark -d. (Note that tshark -d uses the "filter name
of the protocol", which only works for multiple dissectors in the same
table for the same underlying protocol when PINOs are used. Dissector
handles with descriptions are sufficient for the GUI. Cf. #7348.)
We might be able to get rid of the list or postpone the sorting.

This speeds up tshark startup using a default profile by 3-5% in
profiling.