kafka: Use XXH32 from the xxHash library if we have it
The initial implementation of Kafka LZ4 compression incorrectly
included the LZ4 frame magic bytes in the checksum calculation.
To work around this, there's an implementation of XXH32 in the
dissector. Use the version from the library (which is probably
better optimized) if we have it, as we're checking for it now.