register: If we don't have a callback, don't use threads
If we don't have an update callback (e.g., we're running TShark),
then we don't need to spawn threads and do the protocol registrations
and handoffs in single worker threads that we wait on, and lock and
unlock a mutex.

This cuts startup time overhead by a few percent on Linux systems;
it probably varies by OS.

(Theoretically it might be nice to split protocol registration into
multiple threads, but that would take some work. A decent amount of
the calculation time is in parallelizable events like strlen, hashes,
and module_check_valid_name. All the actual inserting into hash tables
and wmem allocation would need to have synchronization primitives,
though.)