epan: Remove a too-late NULL check
proto_register_field_common checks if proto is NULL, but proto has
already been deferenced in proto_register_field_array (and was
similarly dereferenced in the now removed NEW_PROTO_TREE_API).

Apparently it hasn't been a problem, but if we want to check if
proto is NULL, we should do it once at the top of proto_register_field_array
and not once each iteration of a potentially large loop.

Profiling and examination of generated assembly shows that many
compilers do not remove this null check.