wmem_map: Allocate items in chunks
For performance, allocate items as an array in chunks equal to the map
capacity (or change in capacity when growing, which is the same as the
existing capacity). This saves on calling malloc many times versus once,
and on splitting the wmem blocks. It also improves the likelihood that
an item will be found in cache without a RAM hit by keeping the item
data memory addresses closer. The hash table still uses a separate
chaining method, but the items are all stored in an array.

We don't have to realloc the items since we just allocate additional
items. We reuse deleted items, though an option would be to just orphan
them and cleanup when the wmem scope is reset - that would limit to
2^32 items over the lifetime of the map avoid reclaiming memory when
items are removed (without replacing the item at that key), but would
be a bit faster by removing a test on insert. (We don't remove items
from the gpa_name_map in proto.c, which is the largest hash table by far.)

This somewhat increases memory allocation by preallocating. It also
makes 2^32 items a hard limit without more changes (but nothing comes
close to needing that many items, and other hash table implementations,
e.g. GLib, don't support that many items either.)