epan: move filter name check inside tmp_fld_check_assert
Move this check inside the dedicate assertion checking static
function. Note this function is expensive in startup instructions,
because it's executed so many times.

Call module_check_valid_name directly instead of another
(non-static) function, to reduce function overhead a bit.

Don't check again a second time in proto_register_field_init
if the name and abbreviation are empty, because that's already
done in tmp_fld_check_assert, which halts with a dissector bug
if not. That eliminates a lot of instructions and extra branches
because this is called 300,000 times.