btatt: Use register_dissector_with_data to store BT GATT UUIDs
Instead of retrieving the UUID from the dissector protocol name with
g_ascii_strtoull, use register_dissector_with_data and store the
UUID as a pointer (storing 0/NULL for the base dissector.)
This also allows the various BT Generic Attribute dissectors to
be registered using protocols in name only instead of full protocols,
which about half the instruction count for each protocol registration,
of which there are 601 (which is why this is by far the most expensive
dissector handoff function). These don't register any fields and are
just used to provide Decode As options in the table, so PINOs are
appropriate. (They don't need to be separately enabled or disabled,
which would in fact have no effect on dissecting the attribute when it
appears with its normal UUID.)

This has no effect on how the dissectors are used in the Decode As table.
It is backwards and forwards compatible with the UAT as the description
hasn't changed, and as PINOs still works exactly the same with tshark -d
as the (PINO) protocol name hasn't changed.