CMIP, P7: Fix OIDs that failed registrattion
Fix some OIDs where add_oid_from_string was used incorrectly (mostly
switching the order of the parameters, and in one case two strings
failing to be separated becoming concatenated.)

The risk of switching the strings is a reason to use the oid function
with a static array of integers. Another is that the function with
a string simply converts the string to an array of integers, which
is fairly expensive.