expert: Define the EI_INIT values as 0.
We initialize to 0, not -1, now to save static initialzation code.

convert-proto-init.py will remove these for static declarations, but
there's some code (e.g., in Lua) that uses the defines for
initialization and might want to still do so. (In fact, we might
want to have something similar for HF initialization inside arrays,
or dynamic allocation.)