GSM_MAP, Z3950: Don't register OIDs twice
If an oid name lookup is added via register_ber_oid_dissector,
don't call oid_add_from_string a second time. Some of these used
a different name for the second registration. In such cases, I used
the last registered name, which is the one it was renamed to, so
that the dissection stayed the same.

The other cases of duplicate registration involve two different
dissectors, although it's still probably worth eliminating that.