ldaneo: fix build on platforms without 128-bit division.
This is C, not C++, so you can't just use types, rather than formal
parameters, for unused parameters in a function definition.  Visual
Studio, for example, complains "error C2055: expected formal parameter
list, not a type list" for that.

Use the parameter names and decorate them with `_U_` to indicate that
they're unused.