XML: Defer field registration
Parse the processing instructions of the DTDs (and the initial doc type
declaration and first element) with libxml2, which is much faster than
the flex based parser. Get enough information to register media types
and protocols as needed, and then defer actual field registration until
later. (The media type registration must happen so that handoffs are
correct; similarly the protocols have to be registered with their
prefixes.) The full field registration still requires the flex based parser,
at least for now, because libxml2 and other APIs don't make it easy to
get information out of a DTD in the structure we want (as opposed to
using a DTD for validation.)

The flex-based parser also supports both files that contain standalone
external DTDs, and files with a structure that resembles an XML document
with an internal DTD, but no actual XML tags (and thus not a valid
XML document.) Continue to handle both types.

This speeds up initialization, because reading the files with flex
was a significant part of startup.

Ping #20508