wmem: Add a wmem_map_destroy function
Calling wmem_free on a wmem_map_t never freed very much, only the
metadata in the map (e.g., a pointer to an array of pointers to
hash table slots, but not the array of pointers itself), and doesn't
work at all if there are wmem callbacks.

Implement a destroy function that is better than doing that. Ideally
this should not be called and its use is indicative of needing a
code rethink.

In that way, it's similar to the wmem_tree_destroy function, which
is actually not used anywhere in the code.