xml: Preload default media types
It is faster to insert the default media types into the map and
have them be replaced if a particular DTD overrides them than to
check if the type has already been placed in the map by a DTD and
insert them only if not.

The latter performs a moderately expensive string hash twice on all
the media types. Replacing the same key in the hash table with a new
value should occur only rarely, and even if it does it's faster than
hashing twice.