collectd: Separate parts parsing from tapping and column setting
Decrypted data is a container of other parts. In order to handle
decrypting it, the collectd dissector will need to increment the
info column summary data and tap data while processing inside the
container, but only set the columns and queue the tapdata once.

Also remove the special handling for a NULL tree; it doesn't work
properly currently for signed and encrypted packets, and it's not
really worth it because it doesn't optimize over what the API does
already with a NULL tree. When decryption and signature checking is
added, those still couldn't be skipped on a NULL tree, because it
would affect the info column summary stats.

Ping #20634