wmem_map: Ensure that the hash multiplication factor is odd
We multiply the hash results by a random integer before right-shifting,
in order to guard against g_direct_hash being used with integer
(or pointer as integer) values whose upper bits are identical, such
as 16-bit port numbers or pointers allocated nearby each other.

However, this random integer should be an odd random integer; the
larger the power of two that evenly divides x, the more the hash
table space collapses and for smaller tables. The effect becomes
noticeable if x becomes divisible by 2^15 (because of the hash table
of the all the header_field_info, which has ~2^18 entries) and
starts to go into denial of service levels when divisible by powers
of two above 2^24 or so.

Ensure that it's odd by adding 1 if even. This produces a uniform
random distribution over the odd 32-bit integers.