xml: Change cleanup routine to shutdown routine
"init" and "cleanup" routines are called when a capture file or
preferences change. "shutdown" routines are called once right before
the epan scope is destroyed. The encoding pattern GRegex is fixed
and doesn't depend on the preferences or the file, so it should be
created once on registration and use a shutdown routine.