maxmind_db: Don't spawn mmdbresolve at startup if the preference is off
The very first time maxmind_db_post_update_cb is called is before all
the preferences and commandline options have been read. (We load all
the UATs then for various reasons.) In that case, wait to start it
when the nameres preferences are applied. Unconditionally indicate
that the nameres preferences need to be applied when initializing
MaxMindDB, so that it is started even if no nameres preferences are
changed (since the default is true, it needs to be started in that
case.)

Fix #20652