ICMP: introduce ICMP data for ICMP echo
Create a proto tree item representing all ICMP echo data, also the data
that may be interpreted as a timestamp. When it is determined the
timestamp is not available add the data as usual, but for consistency,
adding the ICMP data item as well.
Therefore for any ICMP echo packet 'icmp.data' represents all bytes in
the ICMP data field, while 'data.data' represents these same bytes
except for the optional timestamp.

Fixes #20329