xml: Avoid copying elements unnecessarily
duplicate_element is moderately expensive because it copies the maps
of the element attributes and the pointer arrays of the element
names, among other things.

Rather than copying the pointer arrays and then freeing the copies
after use, just use the original pointer array, and iterate through it.

When constructing a flat namesapce, the xml_ns_t for each element
is used once, on the root_element. Rather than looking each element
up in the map of all elements, copying it, and then later freeing
each original element in the map of all elements, just remove them
as they are used.

When constructing a hiearchical namespace, we can use the original
element for the first level of recursion, and only do the expensive
copies on subsequent levels.