bacpp: Friendly object decoding
Add the objectIdentifier and deviceIdentifiers to the tree and use a
custom format function to provide an abbreviated string, with the type
and instance number consitutent parts below as a bit field.

Note that filtering via string does work with a CF_FUNC, e.g.
bacapp.objectIdentifier == "AV-8052" produces the expected result.
It can also be filtered via the raw 32 bit integer representation.

The abbreviated forms could possibly be used elsewhere, though it's
probably useful to have the long forms somewhere, so as to fully
describe the Reserved and Vendor Proprietary values.

Ping #20025