Aggregation View
For bus systems as CAN, LIN, FlexRay and others messages transport signals,
which typically follow a "last is best" semantics.

For this reason it common to have a live view that updates with the
last messages to see the latest values of the signals.

This patch implements such an aggregation view for live capturing.

Signed-off-by: Dr. Lars Völker <lars.voelker@technica-engineering.de>