MySQL: Properly handle CLIENT_PLUGIN_AUTH_LENENC_CLIENT_DATA
If CLIENT_PLUGIN_AUTH_LENENC_CLIENT_DATA (MYSQL_CAPS_AL) is set,
then the password/authentication data in Login and Change User
packets is a MySQL (variable) length encoded string, not a string
with a fixed one octet with the length. This only matters for
authentication data > 250 octets, and should be rare in practice.

https://dev.mysql.com/doc/dev/mysql-server/latest/page_protocol_connection_phase_packets_protocol_handshake_response.html
https://mariadb.com/docs/server/reference/clientserver-protocol/1-connecting/connection#handshake-response-packet