MySQL: Add preferences for assumed capabilities
Several negotiated capabilities present backwards compatibility
problems, as the MySQL protocol does not provide a way to determine
if they are used or not. The DEPRECATE_EOF capability is a particularly
common offender in this regard. If a capture does not include the
handshake with capability negotiation, dissection will be incorrect and
result in malformed packets.

Add preferences to control whether those capabilities are assumed.
Fix #19856.