print: Don't use hexadecimal for JSON values in -T ek
ElasticSearch and OpenSearch don't automatically convert hexadecimal
representations to integers, even when the mapping type is explicitly
supported. That's more restrictive than many JSON parsers, e.g.
JavaScript's parseInt().

Fix this by introducing a slightly different FTREPR_EK representation.
We might still want to avoid using hexadecimal representation for
integers for the general JSON output, but then again there might
be users who don't convert the values to integers and want to use
them as strings. (The JSON output format needs to be defined better,
possibly distinguishing between "the value of the field in an
appropriate type" and "the preferred string representation of the
field's value.)

For -T ek, we provide a mapping file and assume that the data is
for ingestion by ES/OS, so this is always an error (though one that
can be worked around with an ingest pipeline using a convert processor.)

Fix #15784