ws802_11 + WirelessFrame: Organize frequencies by bands
Not all 802.11 bands support all channel bandwidths. Create separate
arrays for each band, storing the supported channel bandwidths for
each band.

Update the WirelessFrame to support this. This means that impossible
channel bandwidths won't be offered for the 2.4 GHz frequencies,
preventing a failure before it starts. This also means that the
channel QComboBox doesn't have 98 displayed entries (with overlapping
channel numbers) simultaneously, only 11, 28, and 59 for the typical
tri-band case. That makes the combobox more usable, particularly on
themes that disable the pop-up scrollbar.

Add support for EHT 320 GHz in the GUI.

We might want to get additional information for individual frequencies
so as to avoid presenting HT40- or HT40+ as options where they're not
(regulatorily) legal.

nl80211, the only current implementation, provides the capabilities
by band anyway. If someone else ever adds support for macOS (#9921, #19316),
the CoreWLAN api provides a channelBand property that should work.